requirements.txt
Django
stripe
pillow
python-dotenv
requests
